Represent manifest for GitHub Apps (#5)
A new struct has been created that represents the manifest for a GitHub
App[^1]. The intention is that the user provides the path to the
manifest, which is then read from the file and parsed. Some fields can
be overwritten by the user, for example the name of the app. Most fields
are read-only, though, since we want to encourage good defaults in the
shared manifest and little customization on-the-fly.

[^1]: https://docs.github.com/en/apps/sharing-github-apps/registering-a-github-app-from-a-manifest#github-app-manifest-parameters