Bootstrap command to register a GitHub App
The initial structure for commands has been created, using the
`register` command as an example. The command will allow users to
register a new GitHub App from a manifest file. At this point, the
command does nothing more than read a manifest file, parse it, and print
it as JSON to the terminal.