Run a background server for callback from GitHub
When a new GitHub App is registered, GitHub sends a POST request to a
callback URL with a temporary code. The code can be exchanged for the
app's private key and secrets.

The `register` command now starts a web server on a background thread
that listens for the POST request and forwards the code to the command's
main function for further processing.