Fix endpoint for registration form
After the changes in #25, which allowed to provide a custom endpoint for
GitHub, the registration form could not be sent successfully anymore.
One reason was that the path was not added correctly to the base URL,
but the bigger problem was that GitHub uses different domains for its
hosted web interface and API. This is in contrast to self-hosted
instances, where the API is served from a different path (`/api`) and
not a different subdomain (`api.github.com`).