Listen on internal IP address (#31)
The embedded web server has been tweaked to launch on an local IP
address to avoid warnings in the browser. Previously, the local IP was
replaced by the domain `localhost` for the registration form. But this
fix hadn't been applied for the callback , which ended up failing due to
warnings in the browser. Instead, we are now listening on a local-only
IP address, which should avoid th warnings in most (if not all) modern
browsers.